[HER-169] Follow-up Delete the content folder
